refactor: Comment 디렉토리 변경 - components로 이동
