+++ frontend/src/pages/Kanban/style.tsx
@@ -5,7 +5,6 @@ import { Center } from '../../styles/common';
 export const Container = styled.div`
     flex-direction: column;
     overflow-x: hidden;
-    width: 100vw;
     align-items: center;
     ${Center}
 `;
